delete or not

+++ src/components/DeleteCategory/DeleteCategory.js
@@ -19,6 +19,10 @@ export default function DeleteCategory({ category }) {
       return;
     }
 
+    if (!window.confirm("Are you sure you want to delete this category?")) {
+      return;
+    }
+
     deleteDoc(doc(db, "categories", category.id));
   }
